pycode/requirements.txt
langchain
openai < 1
python-dotenv
